added authorisation for adding/updating/deleting product
added authorisation for adding/updating/deleting consultation
added authorisation for updating password/details of user

+++ Files/product/utils.py
@@ -1,8 +1,7 @@
-from unicodedata import category
 from flask_restful import reqparse
 from flask import request, jsonify
 from Files import db
-from ..models import Product, ProductSchema, BelongsToCategory, BelongsToCategorySchema, Seller
+from ..models import User, UserSchema,Product, ProductSchema, BelongsToCategory, BelongsToCategorySchema
 
 
 add_product_args = reqparse.RequestParser()
@@ -15,6 +14,19 @@ add_product_args.add_argument('qty_left', type=int, required=True, help='Product
 add_product_args.add_argument('category', type=int, required=True, help='Product category cannot be blank!')
 add_product_args.add_argument('related_products', type=str, required=True, help='Product related products cannot be blank!')
 
+def check_product_seller_relation(product_id, seller_id):
+    result=db.session.query(Product).filter(Product.product_id==product_id).filter(Product.seller_id==seller_id).first()
+    if not result:
+        return False
+    return True
+
+def check_is_seller(seller_id):
+    result=db.session.query(User).filter(User.user_id==seller_id).first()
+    user_schema=UserSchema()
+    output = user_schema.dump(result)
+    if output["is_seller"] == True:
+        return True
+    return False    
 
 def get_all_products():
     result=Product.query.all()
@@ -23,7 +35,10 @@ def get_all_products():
     return output
 
 def products_by_category(category_name):
-    records = db.session.query(BelongsToCategory).filter(BelongsToCategory.category_name==category_name).filter(BelongsToCategory.pro_con_id!=None).all()
+    records = db.session.query(BelongsToCategory).filter(
+        BelongsToCategory.category_name==category_name).filter(
+            BelongsToCategory.pro_con_id!=None).filter(
+                BelongsToCategory.pro_con_id.startswith('P')).all()
     result = []
     for record in records:
         output = jsonify(
@@ -52,12 +67,12 @@ def get_product_by_id(id):
 
 def add_product(name, description, price, image, discount, qty_left, categories, related_products, seller_id):
     
-    seller_exists=db.session.query(Seller).filter(Seller.seller_id==seller_id).first()
-    if not seller_exists:
-        return {"message": "Seller doesn't exists"}
+    if check_is_seller(seller_id) == False:
+        return {"message": "You are not a seller"}, 400
+
     result = db.session.query(Product).filter(Product.name==name).filter(Product.seller_id==seller_id).first()
     if result:
-        return {"message":"Product Already Exists"}
+        return {"message":"Product Already Exists"}, 409
     #1. Adding Product
     record=Product(name=name,description=description,price=price,image=image,
                     discount=discount,effective_price=float(price)-(float(discount)*float(price)/100),
@@ -66,8 +81,7 @@ def add_product(name, description, price, image, discount, qty_left, categories,
     
     #2. Getting Prod ID
     
-    temp = db.session.query(Product).filter(Product.name==
-                                                     name).filter(Product.seller_id==seller_id).first()
+    temp = db.session.query(Product).filter(Product.name==name).filter(Product.seller_id==seller_id).first()
     output = ProductSchema(many=False).dump(temp)
     product_id = jsonify(output).json["product_id"]
     product_id = "P" + str(product_id)
@@ -85,15 +99,23 @@ def add_product(name, description, price, image, discount, qty_left, categories,
                 db.session.add(BelongsTo)
             else:
                 return {"message": "Wrong Category Entered"}, 400
+
     db.session.commit()
     
     db.session.commit()
     return {"message": "Done"}, 201
 
-def update_product(product_id, name, description, price, image, discount, qty_left, categories, related_products):
+def update_product(product_id, name, description, price, image, discount, qty_left, categories, related_products, seller_id):
+    
+    if check_is_seller(seller_id) == False:
+        return {"message": "You are not a seller"}, 400
+
+    if check_product_seller_relation(product_id, seller_id) == False:
+        return {"message": "You are not the seller of this product"}, 400
+
     product=db.session.query(Product).filter(Product.product_id==product_id).first()
     if not product:
-        return {"message": "Product does not exist"}
+        return {"message": "Product does not exist"}, 204
     
     #1. Update product table
     product.name=name
@@ -112,7 +134,7 @@ def update_product(product_id, name, description, price, image, discount, qty_le
         db.session.delete(record)
     
     # 3. Adding new Product Category Mapping
-    consultation_id = "P" + str(consultation_id)
+    product_id = "P" + str(product_id)
     for CategoryName in categories.split(","):
         temp = db.session.query(BelongsToCategory).filter(BelongsToCategory.category_name == CategoryName).first()
         if not temp is None:
@@ -120,7 +142,7 @@ def update_product(product_id, name, description, price, image, discount, qty_le
                 BelongsToCategorySchema(many=False).dump(temp)
                 )
             BelongsTo = BelongsToCategory(category_name = output.json["category_name"], 
-                                        pro_con_id = consultation_id)
+                                        pro_con_id = product_id)
             db.session.add(BelongsTo)
         else:
             return {"message": "Product Modified but Wrong Category(s) Entered"}, 400
@@ -128,10 +150,21 @@ def update_product(product_id, name, description, price, image, discount, qty_le
 
     return {"message": "Done"}, 202
 
-def remove_product(id):
-    product=db.session.query(Product).filter(Product.product_id==id).first()
+def remove_product(product_id, seller_id):
+
+    if check_is_seller(seller_id) == False:
+        return {"message": "You are not a seller"}, 400
+
+    if check_product_seller_relation(product_id, seller_id) == False:
+        return {"message": "You are not the seller of this product. You cannot delete this product"}, 400
+
+    product=db.session.query(Product).filter(Product.product_id==product_id).first()
     if not product:
         return None
+    product_id = "P"+str(product_id)
+    records = db.session.query(BelongsToCategory).filter(BelongsToCategory.pro_con_id==product_id).all()
+    for record in records:
+        db.session.delete(record)
     db.session.delete(product)
     db.session.commit()
-    return {"message": "Done"}, 204
+    return {"message": "Done"}
